# ENH speed up test_unsupervised_model_fit