--- a/sklearn/feature_selection/tests/test_sequential.py
+++ b/sklearn/feature_selection/tests/test_sequential.py
@@ -148,12 +148,12 @@ def test_pipeline_support():
     pipe.transform(X)
 
 
-@pytest.mark.parametrize("n_features_to_select", (2, 3, 4))
+@pytest.mark.parametrize("n_features_to_select", (2, 3))
 def test_unsupervised_model_fit(n_features_to_select):
     # Make sure that models without classification labels are not being
     # validated
 
-    X, y = make_blobs(n_features=6)
+    X, y = make_blobs(n_features=4)
     sfs = SequentialFeatureSelector(
         KMeans(),
         n_features_to_select=n_features_to_select,