<!--
Thanks for contributing a pull request! Please ensure you have taken a look at
the contribution guidelines: https://github.com/scikit-learn/scikit-learn/blob/main/CONTRIBUTING.md
-->

#### Reference Issues/PRs
<!--
Example: Fixes #1234. See also #3456.
Please use keywords (e.g., Fixes) to create link to the issues or pull requests
you resolved, so that they will automatically be closed when your pull request
is merged. See https://github.com/blog/1506-closing-issues-via-pull-requests
-->
Resolving issue #21407 . 

**Detail:**
Changed test parameters from [2,3,4] to [2,3], and reduced feature dimension from 6 to 4. Resulting in a ~60% reduction in run time on my machine.

Before: 
```
1.14s call     sklearn/feature_selection/tests/test_sequential.py::test_unsupervised_model_fit[4]
0.97s call     sklearn/feature_selection/tests/test_sequential.py::test_unsupervised_model_fit[3]
0.71s call     sklearn/feature_selection/tests/test_sequential.py::test_unsupervised_model_fit[2]
```

After:
```
0.59s call     sklearn/feature_selection/tests/test_sequential.py::test_unsupervised_model_fit[3]
0.49s call     sklearn/feature_selection/tests/test_sequential.py::test_unsupervised_model_fit[2]
```

<!--
Please be aware that we are a loose team of volunteers so patience is
necessary; assistance handling other issues is very welcome. We value
all user contributions, no matter how minor they are. If we are slow to
review, either the pull request needs some benchmarking, tinkering,
convincing, etc. or more likely the reviewers are simply busy. In either
case, we ask for your understanding during the review process.
For more information, see our FAQ on this topic:
http://scikit-learn.org/dev/faq.html#why-is-my-pull-request-not-getting-any-attention.

Thanks for contributing!
-->